Add support for loading images with a device scale
Images can now be loaded assuming a device scale for a given image
file (with the current `ply_image_load` assuming a device scale of 1).
Alternatively, the code can look for images with an `@` suffix in their
name containing the device scale the file was meant for. This can either
be done with or without the default name at scale 1 as a fallback.

Also add a utility for splitting a file path into the part up to the dot
of the last file name extension and the extension, including the dot.