x11: Support function keys for keyboard input
Plymouth now supports getting function key input from
the tty and passing it on to plugins and scripts.

The same support doesn't currently exist when key input
comes from an X server.

This commit adds support for translating function key
events from the X server to something the reset of
plymouth can understand.

Note, this is only handles functions keys, but there
are other sequences that could be translated as well.

The types of keys supported may grow in the future.

Fixes #166