ply-input-device: Add new plymouth.debug-input-devices kernel option
Right now we don't have a lot of debugging information when processing
keyboard events.  I mean we don't want passwords to show up in logs
after all.

Still, it would be useful to have a way to turn the valve on to the
firehose.

This commit adds a plymouth.debug-input-devices kernel option.