main: Add new "presentation" mode
In some kiosk scenarios, the running service doesn't need a full fledged
display server, but instead could get away with plymouth in a sort of
"billboard" mode, where it lets the user know the service is running,
but doesn't need to give any sort of indication on progress.

This commit adds a new mode for this purpose (called "presentation"
mode). In order to provide backward compatibility, it immediately
tells the splash plugin to become idle once being displayed. For
many cases, this should be a good enough experience out of the box.

(and other cases can just extend their splash plugin to have specific
support for the presentation mode).