ply-key-file: Introduce ply_key_file_get_color
It acts similar to ply_key_file_get_long except now, when the value
is preceded by a colon, it pads the alpha values differently. Instead
of ANDing 0xff000000, it determines if the specified value was specified
with 6 bytes or not. This allows values with no red values to be handled
correctly