scripts: include label-pango.so in initrd
Debians dracut package uses plymouth-populate-initrd to include plymouth
in the built initrd. However, the label-pango.so module won't be
included by that script, meaning that prompts won't show any text labels
for the default theme on Debian (emerald, from the desktop-base
package, the reason this isn't a bigger issue is that Debian defaults to
using initramfs-tools, and not dracut).

Including label-pango.so in addition to label-freetype.so fixes that and
shouldn't have any ill effects on other platforms.